:recycle: refactor : rename packages global.exceptions and global.utils to singular
