+++ src/main/java/com/run_us/server/domains/running/exception/RunningErrorCode.java
@@ -1,6 +1,6 @@
 package com.run_us.server.domains.running.exception;
 
-import com.run_us.server.global.exceptions.enums.CustomResponseCode;
+import com.run_us.server.global.exception.enums.CustomResponseCode;
 import org.springframework.http.HttpStatus;
 
 public enum RunningErrorCode implements CustomResponseCode {
