+++ src/main/java/com/run_us/server/domains/running/exception/RunningException.java
@@ -1,7 +1,7 @@
 package com.run_us.server.domains.running.exception;
 
-import com.run_us.server.global.exceptions.BusinessException;
-import com.run_us.server.global.exceptions.enums.CustomResponseCode;
+import com.run_us.server.global.exception.BusinessException;
+import com.run_us.server.global.exception.enums.CustomResponseCode;
 
 public class RunningException extends BusinessException {
