+++ src/main/java/com/run_us/server/global/common/Response.java
@@ -1,6 +1,6 @@
 package com.run_us.server.global.common;
 
-import com.run_us.server.global.exceptions.enums.CustomResponseCode;
+import com.run_us.server.global.exception.enums.CustomResponseCode;
 import lombok.Getter;
 
 @Getter
