+++ src/main/java/com/run_us/server/global/exception/BusinessException.java
@@ -1,6 +1,6 @@
-package com.run_us.server.global.exceptions;
+package com.run_us.server.global.exception;
 
-import com.run_us.server.global.exceptions.enums.CustomResponseCode;
+import com.run_us.server.global.exception.enums.CustomResponseCode;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
