+++ src/main/java/com/run_us/server/global/exception/StompErrorHandler.java
@@ -1,4 +1,4 @@
-package com.run_us.server.global.exceptions;
+package com.run_us.server.global.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Configuration;
