+++ src/main/java/com/run_us/server/global/exception/enums/ExampleErrorCode.java
@@ -1,4 +1,4 @@
-package com.run_us.server.global.exceptions.enums;
+package com.run_us.server.global.exception.enums;
 
 import org.springframework.http.HttpStatus;
