+++ src/main/java/com/run_us/server/global/util/PassCodeGenerator.java
@@ -1,4 +1,4 @@
-package com.run_us.server.global.utils;
+package com.run_us.server.global.util;
 
 import java.security.SecureRandom;
 import java.util.Random;
